# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Identity API",
    "description": "API for the my identity service.",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "identifier": "Apache-2.0"
    },
    "contact": {
      "name": "Trent Shailer"
    }
  },
  "servers": [
    {
      "url": "https://identity.trentshailer.com/v1"
    }
  ],
  "paths": {
...
    "/public-keys/{publicKeyId}": {
      "parameters": [
        {
          "name": "publicKeyId",
          "description": "The `rawId` of the public key.",
          "required": true,
          "in": "path",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "description": "Get the specified public key.",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicKey"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "description": "Delete the specified public key.",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "description": "Create a new session for the specified public key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicKeyCredentialAssertion"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "token": []
    }
  ],
  "components": {
    "responses": {
      "400": {
        "description": "Invalid Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "401": {
        "description": "Unauthenticated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "problems"
        ],
        "properties": {
          "problems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem": {
        "type": "object",
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "type": "string"
          },
          "pointer": {
            "description": "A JSON path that identifies the part of the request that was the cause of the problem.",
            "type": "string"
          }
        }
      },
...
      "PublicKey": {
        "type": "object",
        "required": [
          "rawId",
          "indentityId",
          "displayName",
          "publicKey",
          "publicKeyAlgorithm",
          "transports",
          "created"
        ],
        "properties": {
          "rawId": {
            "type": "string"
          },
          "identityId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "publicKey": {
            "type": "string",
            "format": "base64"
          },
          "publicKeyAlgorithm": {
            "type": "integer"
          },
          "transports": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created": {
            "type": "string",
            "format": "rfc3339"
          },
          "lastUsed": {
            "type": "string",
            "format": "rfc3339"
          }
        }
      },
...
      "PublicKeyCredentialAssertion": {
        "type": "object",
        "required": [
          "id",
          "rawId",
          "type",
          "response"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "rawId": {
            "type": "string",
            "format": "base64"
          },
          "authenticatorAttachment": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "response": {
            "type": "object",
            "required": [
              "authenticatorData",
              "clientDataJSON",
              "signature",
              "userHandle"
            ],
            "properties": {
              "authenticatorData": {
                "type": "string",
                "format": "base64"
              },
              "clientDataJSON": {
                "type": "string",
                "format": "base64"
              },
              "signature": {
                "type": "string",
                "format": "base64"
              },
              "userHandle": {
                "type": "string",
                "format": "base64"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-TS-API-Key",
        "in": "header"
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}